+++ packages/target-tauri/src-tauri/src/lib.rs
@@ -13,6 +13,7 @@ mod i18n;
 mod runtime_info;
 mod settings;
 mod state;
+mod stickers;
 mod temp_file;
 mod webxdc;
 
@@ -116,6 +117,7 @@ pub fn run() {
         ])
         .register_asynchronous_uri_scheme_protocol("webxdc-icon", webxdc::webxdc_icon_protocol)
         .register_asynchronous_uri_scheme_protocol("dcblob", blobs::delta_blobs_protocol)
+        .register_asynchronous_uri_scheme_protocol("dcsticker", stickers::delta_stickers_protocol)
         .setup(move |app| {
             // Create missing directories for iOS (quick fix, better fix this upstream in tauri)
             #[cfg(target_os = "ios")]